+++ src/std/mod.rs
@@ -8,9 +8,9 @@ mod unix;
 mod windows;
 
 #[cfg(target_os = "windows")]
-pub use self::windows::tx_rx_task;
+pub use self::windows::{ethercat_now, tx_rx_task};
 #[cfg(unix)]
-pub use unix::tx_rx_task;
+pub use unix::{ethercat_now, tx_rx_task};
 // io_uring is Linux-only
 #[cfg(target_os = "linux")]
 pub use io_uring::tx_rx_task_io_uring;