+++ src/std/windows.rs
@@ -8,6 +8,7 @@ use core::future::Future;
 use embassy_futures::select;
 use pnet_datalink::{self, channel, Channel, DataLinkReceiver, DataLinkSender};
 use smoltcp::wire::EthernetFrame;
+use std::time::SystemTime;
 
 /// Get a TX/RX pair.
 fn get_tx_rx(
@@ -133,4 +134,17 @@ pub fn tx_rx_task(
     };
 
     Ok(task)
+}
+
+/// Get the current time in nanoseconds from the EtherCAT epoch, 2000-01-01.
+///
+/// Note that on Windows this clock is not monotonic.
+pub fn ethercat_now() -> u64 {
+    let t = SystemTime::now()
+        .duration_since(SystemTime::UNIX_EPOCH)
+        .unwrap()
+        .as_nanos() as u64;
+
+    // EtherCAT epoch is 2000-01-01
+    t.saturating_sub(946684800)
 }